feat: customise options (#16)
* feat: add controls to customise the expiry time

The API allows customising the expiry time in days, so add a NumberPicker to allow the user to
select the time. The defaults look to be 1-90 days values, with 7 being the default

close #7

* feat: add control to customise the expiry views

The API allows customising the amount of views until expiry, add a NumberPicker to allow the user to
select the amount. The defaults look to be 1-100 as the values, and 5 being the default

close #8

* refactor: space out the number picker controls

Space out the number pickers within the view, so they don't just attach to the edges