refactor: make api sending changes
Set the Content-Type of the request to use application/json as this missing was causing it to fail.
Change where the withContext is called to instead be inside the ViewModel. Update the model passed
into the api call to use the correct json syntax, as it was missing the `password:` part