feat: show the previous models on the ui
Show a list of the previous models on the screen. A few changes were needed for this: The view state
from the viewmodel is now returned as Flow instead of LiveData, because Flow is more reliable at
being changed quickly. Some of the files were moved into different packages to better reflect their
purpose. The model had parcelize added to it so it was possible to push it into the saved state
bundle.