feat: show previous models in ui (#324)
* feat: show the previous models on the ui

Show a list of the previous models on the screen. A few changes were needed for this: The view state
from the viewmodel is now returned as Flow instead of LiveData, because Flow is more reliable at
being changed quickly. Some of the files were moved into different packages to better reflect their
purpose. The model had parcelize added to it so it was possible to push it into the saved state
bundle.

* feat: add tap to copy previous url

Add ability to tap on a previous model and the url will be copied to the clipboard.

* feat: add new domain object for displaying pushed models

Add new domain so we can keep the remote data source data and pusher domain data seperate. With this
the date is also now being formatted to be displayed nicely to the user.

* feat: add preference button to clear pushed history

Add button in the preferences screen which allows the user to clear all of their pushed history in
one go.

* fix: ja strings in wrong location

JA strings were nested a folder too deep and weren't working

* docs: add missing kdoc on domain object

* docs: fix typo

* refactor: change date string formatting to its own class

Move the formatting of the date string into its own class, and have it format the date into the
users chosen format, instead of a specified format by the app.

* style: fix some detekt warnings